+++ changelog.md
@@ -4,7 +4,7 @@ Changelog for npm and snapcraft
 
 ## Version 1.1.0
 
-- Colors can now be customized in `.taskline.json`. Possible are all foreground colors of [chalk](https://github.com/chalk/chalk#colors).
+- Colors can now be customized in `.taskline.json`. Possible are all foreground colors of [chalk](https://github.com/chalk/chalk#colors), rgb and hex colors.
 
 ```json
 "theme": {